Deletar Produtos

+++ src/app/modules/shared/components/confirm/confirm.component.ts
@@ -1,6 +1,7 @@
 import { Component, inject, OnInit } from '@angular/core';
 import { CategoryService } from '../../services/category.service';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { ProductService } from '../../services/product.service';
 
 @Component({
   selector: 'app-confirm',
@@ -10,6 +11,7 @@ import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 export class ConfirmComponent implements OnInit {
 
   private categoryService = inject(CategoryService);
+  private productService = inject(ProductService);
   private dialogRef = inject(MatDialogRef)
   public data = inject(MAT_DIALOG_DATA);
 
@@ -18,19 +20,29 @@ export class ConfirmComponent implements OnInit {
   onConfirm() {
     // Lógica para confirmação
     if (this.data != null) {
-      this.categoryService.deleteCategory(this.data.id)
-      .subscribe((data: any) => {
-        this.dialogRef.close(1);
-      }, (error: any) => {
-        this.dialogRef.close(2);
-      });
+
+      if (this.data.module == 'category') {
+
+        this.categoryService.deleteCategory(this.data.id)
+        .subscribe((data: any) => {
+          this.dialogRef.close(1);
+        }, (error: any) => {
+          this.dialogRef.close(2);
+        });
+    } else if (this.data.module == 'product') {
+        this.productService.deleteProduct(this.data.id)
+        .subscribe((data: any) => {
+          this.dialogRef.close(1);
+        }, (error: any) => {
+          this.dialogRef.close(2);
+        });
     }else{
       this.dialogRef.close(2);
   }
 }
+  }
   onCancel() {
     // Lógica para cancelamento
     this.dialogRef.close(3);
   }
-
 }
